Separate checkout analytics integration test
Committed-By-Agent: codex
